refactor unit tests and add one to cover list_datasets
